silence some new clang-tidy warnings

+++ src/joedb/journal/SHA_256.h
@@ -66,7 +66,7 @@ namespace joedb
   public:
    SHA_256(): hash(hash_init) {}
    const Hash &get_hash() const {return hash;}
-   static constexpr int chunk_size = 64;
+   static constexpr size_t chunk_size = 64;
 
    /////////////////////////////////////////////////////////////////////////
    void process_chunk(const char *data)
